requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCl
dnspython
discord.py
async-timeout